+++ UA Core Library.sln
@@ -1,4 +1,3 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30517.126
@@ -28,6 +27,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 		azure-pipelines-preview.yml = azure-pipelines-preview.yml
 		azure-pipelines.yml = azure-pipelines.yml
+		Tests\benchmarks.cmd = Tests\benchmarks.cmd
 		Tests\codecoverage.cmd = Tests\codecoverage.cmd
 		Tests\codecoverage.sh = Tests\codecoverage.sh
 		common.props = common.props
